WIP Merge into main

+++ ESP32/DIY-Flow-Bench/mafData/ACDELCO_19330122.cpp
@@ -68,7 +68,7 @@ int Maf::mafUnits() {
  * @brief mafLookupTable
  * @note Global vector of vectors containing MAF>Flow key>value pairs
  ***/
- std::vector<std::vector<int>> mafLookupTable = {{ 
+ std::vector<std::vector<long>> mafLookupTable = {{ 
 		{0,0},
 		{1500,142},
 		{1625,176},
