+++ ESP32/DIY-Flow-Bench/mafData/SIEMENS_5WK9605.cpp
@@ -69,7 +69,7 @@ int Maf::mafUnits() {
  * @brief mafLookupTable
  * @note Global vector of vectors containing MAF>Flow key>value pairs
  ***/
- std::vector<std::vector<int>> mafLookupTable = {{ 
+ std::vector<std::vector<long>> mafLookupTable = {{ 
 {0,0},
 {100,6226},
 {200,6745},
